monitoring-host/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
django-storages==1.14.2  # For S3 storage
boto3==1.34.34  # For S3 storage
python-dotenv==1.0.0
Pillow==10.2.0  # For image handling
django-filter==23.5
markdown==3.5.1 